tests(refactor): Modify file directory
