--- tests/iot/query.go
@@ -1,1 +1,0 @@
-package iot
