+++ tests/suite/iot/query.go
@@ -1,0 +1,1 @@
+package iot
